chore: add `hex_to_rgb` as a shared function
This function is duplicated across several apps. By making this a shared
function now we can migrate other apps such as Last.fm and clock to use
the shared function in the future.
