+++ repository.txt
@@ -17,6 +17,7 @@
     "lib/font_large.py",
     "lib/ai_agent.py",
     "lib/cmd.py",
-    "lib/filemanager.py"
+    "lib/filemanager.py",
+    "lib/colors.py"
   ]
 }
